Grapple V.1
First Grapple Version. This version is fully working but some tweaks on how it interacts with certain elements would be needed. I encountered a bug when adding upward movement at the end of the hook when hitting a wall to help the player climb a wall so I left it out of this version.