Started first spell
The bug has defeated me. I dont know how to fix it I have tired so much. I moved on to the spell and converted the projectile that the gun shoots to be fire. I also made it so the fireball destroys other physics objects when it hits them. I dont know how to change the mesh of the gun in the players hands might work on tomorow. Also changed the weapon pick up to be a book.